refactor: group DI modules under di/ package
Mirrors the GetIcon reference project's convention of keeping Hilt
modules and qualifiers in a dedicated di package instead of the root
package.
