+++ app/src/main/java/de/lemke/oneuisample/di/PersistenceModule.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.lemke.oneuisample
+package de.lemke.oneuisample.di
 
 import android.content.Context
 import android.content.Context.MODE_PRIVATE
