+++ app/src/test/java/de/lemke/oneuisample/di/TestDispatchersModule.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.lemke.oneuisample
+package de.lemke.oneuisample.di
 
 import dagger.Module
 import dagger.Provides
